refactor: moved useMail to hooks folder and updated imports
