+++ app/components/Contacts/Contacts.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { PaperAirplaneIcon } from '@heroicons/react/24/outline'
 import { Button, Typography } from '@/app/components/ui'
-import { useMail } from '../Greetings/useMail'
+import { useMail } from '@/app/hooks/useMail'
 
 export default function Contacts() {
   const { emailMeHref } = useMail()
@@ -13,8 +13,7 @@ export default function Contacts() {
       <div className="flex flex-wrap items-center justify-between gap-4 lg:flex-nowrap">
         <article className="w-full text-xs md:w-8/12">
           <Typography.Description>
-            Lorem Ipsum is simply dummy text of the printing and typesetting
-            industry.
+          I'm happy to have you here! Let's connect and make some remarkable project together!
           </Typography.Description>
         </article>
         <div className="flex max-h-[36px] w-full justify-end gap-2 md:w-fit">
