+++ app/components/ui/index.ts
@@ -1,5 +1,6 @@
 import { Button } from '@/app/components/ui/Button/Button'
+import { ButtonClient } from '@/app/components/ui/Button/ButtonClient'
 import { Typography } from '@/app/components/ui/Typography/Typography'
 import { Tags } from '@/app/components/ui/Tags/Tags'
 
-export { Button, Typography, Tags }
+export { Button, ButtonClient, Typography, Tags }
